% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_until.R
\name{run_until}
\alias{run_until}
\title{run_until}
\usage{
run_until(FUN, secs = 0)
}
\arguments{
\item{FUN}{function.}

\item{secs}{seconds.}
}
\description{
Keep running a function until a certain time.
}
\examples{
run_until(stop(), secs = 5)
}
\author{
Charco Hui
}
