% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solidArrow.R
\name{solidArrow}
\alias{solidArrow}
\title{solidArrow}
\usage{
solidArrow(x0, y0, x1, y1, angle = 30, col = "black", ...)
}
\arguments{
\item{x0, x1}{function.}

\item{angle}{Angle from the shaft of the arrow to the edge of the arrow head.}
}
\description{
Draw a arrow with solid arrow head
}
\examples{
plot(1)
solidArrow(0, 0, 1, 1)
}
\author{
Charco Hui
}
